requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
psycopg2-binary  # For connecting to PostgreSQL
python-dotenv    # To manage environment variables from .env file
gunicorn         # Production web server for Render
pandas           # For data analysis and CSV reports
fpdf             # For PDF reports
werkzeug         # For password hashing